requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.3
django-extensions==3.2.3
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pip==23.3.1
platformdirs==4.2.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.0
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
setuptools==68.2.2
sqlparse==0.4.4
urllib3==2.2.1
wheel==0.41.2
yarl==1.9.4
